% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial_Distribution.R
\name{BetaBinomial}
\alias{BetaBinomial}
\title{Constructor for Beta Binomial distribution}
\usage{
BetaBinomial(model_object, outcome, group = NA, opt = NA)
}
\arguments{
\item{model_object}{an object coming from the \code{fit_model()} function}

\item{outcome}{the variable to be used as the dependent variable provided
as an expression}

\item{group}{the variable to be used as the grouping variable provided
as an expression}

\item{opt}{the model options provided as a function call to \code{model_options()}}
}
\value{
An object of class BetaBinomial
}
\description{
Constructor for Beta Binomial distribution
}
