% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial_Distribution.R
\name{Binomial}
\alias{Binomial}
\title{Constructor for Binomial distribution}
\usage{
Binomial(model_object, outcome, group = NA, opt = NA)
}
\arguments{
\item{model_object}{an object coming from the \code{fit_model()} function}

\item{outcome}{the variable to be used as the dependent variable provided
as an expression}

\item{group}{the variable to be used as the grouping variable provided
as an expression}

\item{opt}{the model options provided as a function call to \code{model_options()}}
}
\value{
An object of class Binomial.
}
\description{
Constructor for Binomial distribution
}
