% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine all individual distribution plots}
\usage{
combine_plots(model_array, hist = FALSE)
}
\arguments{
\item{model_array}{a tibble retreived from the \code{fit_model()} function}

\item{hist}{A logical value indicating whether a histogram of the data
should be overlayed on the graph}
}
\value{
A ggplot object containing the overlayed graphs.
}
\description{
\code{combine_plots()} combines each plot and overlay them onto one graph.
}
\details{
If a grouping varaible was specified in \code{fit_model()} one graph is made for
each group.
}
