% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{get_predictors}
\alias{get_predictors}
\title{Returns a vector of predictors}
\usage{
get_predictors(model_eq)
}
\arguments{
\item{model_eq}{A model equation as a string}
}
\value{
A vector with each variable found in model_eq as a string
}
\description{
\code{get_predictors()} returns a list with the variable names found in the model equation.
}
\examples{
get_predictors("y~1")
get_predictors("y~height + weight")

}
