% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial_Distribution.R
\name{make_betabinom}
\alias{make_betabinom}
\title{Make Beta Binomial Distribution}
\usage{
make_betabinom(...)
}
\arguments{
\item{...}{A comma separated list of predictors that will be used in the
model, given as unquoted expressions. 1 if fitting an intercept
only model}
}
\value{
A tibble with one row. The tibble will contain information stating
that it is a beta binomial model and will contain the right hand side of
a model formula by creating a linear combination of predictors
}
\description{
\code{make_betabinom()} creates an object that allows for analysis assuming
a binomial distribution.
}
\details{
The returned tibble can be row bound with other objects from the
\code{make_distribution()} functions. Then multiple models can be worked
with at once.

This tibble is the input for \code{set_priors()} and \code{fit_model()} functions.
}
\examples{
make_betabinom(1)
make_betabinom(x)
make_betabinom(x, y)

}
