% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{make_distribution}
\alias{make_distribution}
\title{Make an arbitrary distribution}
\usage{
make_distribution(dist, ...)
}
\arguments{
\item{dist}{an expression containing the name of the distribution}

\item{...}{a comma separated list of predictors that will be in the
model, given as unquoted expressions}
}
\value{
A tibble with one row. The tibble will contain information stating
the name of the distribution and the right hand side of the model
formula by forming a linear combination of the given predictors.
}
\description{
\code{make_distribution()} returns an object that represents a distribution
}
\details{
This is a generic function which takes in a distribution name and a set
comma separated predictiors and returns a tibble with the information.
}
\examples{
make_distribution(binomial, 1)
make_distribution(normal, weight, height)
}
