% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{model_options}
\alias{model_options}
\title{Set options for a model}
\usage{
model_options(...)
}
\arguments{
\item{...}{A list of unquoted expressions setting the options in the form
of \code{option = value}.}
}
\value{
Returns a named list containing the model options specified
}
\description{
\code{model_options()} sets model specific options for use within \code{model_fit()}.
}
