% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot each distribution}
\usage{
plot_distribution(model_array, hist = FALSE)
}
\arguments{
\item{model_array}{a tibble retreived from the \code{fit_model()} function}

\item{hist}{a logical indicating whether a histogram should overlay the
distribution}
}
\value{
A tibble with each row representing one distribution. The \code{graphics}
column contains a ggplot object with the specific graph. Additional
model information is returned including the type of model and the
equation corresponding to the model.
}
\description{
\code{plot_distribution()} plots each distribution on a separate graph and returns
a tibble with the graphs.
}
\details{
If a grouping variable was supplied to \code{fit_model()} a graph is made for
each group.
}
