% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{set_priors}
\alias{set_priors}
\title{Assigns predictors to a model}
\usage{
set_priors(model_array, ...)
}
\arguments{
\item{model_array}{a tibble containing the model information (distribution
and model equation). Often the output of a \code{make_distribution()}
function.}

\item{...}{a list of expressions set up as "predictor = distribution"
where the distribution is written in JAGS syntax}
}
\value{
a modified version of model_array that has the relevant predictors
for each model(row)
}
\description{
\code{set_priors()} sets the priors for each variable in a model.
}
\details{
Each prior is written following JAGS notation and specified using
\code{variable = JAGS}. User must specify a prior for the intercept
using \code{int=}. The priors are model specific so if the user
supplies priors for variables that are not found in the model
equation they will not be retained.
}
